+++ crates/lib/core/tests/math/u64_mod.rs
@@ -4,7 +4,7 @@ use miden_core::assert_matches;
 use miden_core_lib::handlers::u64_div::{U64_DIV_EVENT_NAME, U64DivError};
 use miden_processor::ExecutionError;
 use miden_utils_testing::{
-    Felt, TRUNCATE_STACK_PROC, U32_BOUND, ZERO, expect_exec_error_matches, proptest::prelude::*,
+    Felt, TRUNCATE_STACK_PROC, U32_BOUND, expect_exec_error_matches, proptest::prelude::*,
     rand::rand_value,
 };
 
@@ -727,11 +727,10 @@ fn checked_and_fail() {
 
     expect_exec_error_matches!(
         test,
-        ExecutionError::NotU32Values{ values, err_code, label: _, source_file: _ } if
+        ExecutionError::NotU32Values{ values, label: _, source_file: _ } if
             values.len() == 2 &&
             values.contains(&Felt::new(a0)) &&
-            values.contains(&Felt::new(b0)) &&
-            err_code == ZERO
+            values.contains(&Felt::new(b0))
     );
 }
 
@@ -773,11 +772,10 @@ fn checked_or_fail() {
 
     expect_exec_error_matches!(
         test,
-        ExecutionError::NotU32Values{ values, err_code, label: _, source_file: _ } if
+        ExecutionError::NotU32Values{ values, label: _, source_file: _ } if
             values.len() == 2 &&
             values.contains(&Felt::new(a0)) &&
-            values.contains(&Felt::new(b0)) &&
-            err_code == ZERO
+            values.contains(&Felt::new(b0))
     );
 }
 
@@ -819,11 +817,10 @@ fn checked_xor_fail() {
 
     expect_exec_error_matches!(
         test,
-        ExecutionError::NotU32Values{ values, err_code, label: _, source_file: _ } if
+        ExecutionError::NotU32Values{ values, label: _, source_file: _ } if
             values.len() == 2 &&
             values.contains(&Felt::new(a0)) &&
-            values.contains(&Felt::new(b0)) &&
-            err_code == ZERO
+            values.contains(&Felt::new(b0))
     );
 }